nn/requirements.txt
keras
tensorflow
pillow
sklearn
